+++ test/conf/cassandra-jmx-pem-sslconfig.yaml
