+++ test/unit/org/apache/cassandra/utils/JMXSslDisabledEncryptionOptionsTest.java
@@ -23,14 +23,12 @@ import java.util.Map;
 import javax.management.remote.rmi.RMIConnectorServer;
 import javax.net.ssl.SSLException;
 
-import org.apache.commons.lang3.StringUtils;
 import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
 import org.apache.cassandra.config.DatabaseDescriptor;
-import org.apache.cassandra.config.EncryptionOptions;
 import org.apache.cassandra.distributed.shared.WithProperties;
 
 import static org.apache.cassandra.config.CassandraRelevantProperties.CASSANDRA_CONFIG;
@@ -38,14 +36,17 @@ import static org.apache.cassandra.config.CassandraRelevantProperties.COM_SUN_MA
 import static org.apache.cassandra.config.CassandraRelevantProperties.JAVAX_RMI_SSL_CLIENT_ENABLED_CIPHER_SUITES;
 import static org.apache.cassandra.config.CassandraRelevantProperties.JAVAX_RMI_SSL_CLIENT_ENABLED_PROTOCOLS;
 
-public class JMXSslPluggableConfigTest
+/**
+ * This class tests for disabling jmx_encryption_options in the cassandra.yaml.
+ */
+public class JMXSslDisabledEncryptionOptionsTest
 {
     static WithProperties properties;
 
     @BeforeClass
     public static void setupDatabaseDescriptor()
     {
-        properties = new WithProperties().set(CASSANDRA_CONFIG, "cassandra-jmx-sslcontextfactory.yaml");
+        properties = new WithProperties().set(CASSANDRA_CONFIG, "cassandra-jmx-disabled-sslconfig.yaml");
         DatabaseDescriptor.daemonInitialization();
     }
 
@@ -55,22 +56,41 @@ public class JMXSslPluggableConfigTest
         properties.close();
     }
 
+    /**
+     * Tests absence of any JMX SSL configuration out of the below three,
+     * 1. local only JMX server
+     * 2. System properties set for remote JMX SSL
+     * 3. jmx_encryption_options in the cassandra.yaml
+     * @throws SSLException
+     */
     @Test
-    public void testJMXPluggableConfig() throws SSLException
+    public void testJMXDisabledEncryptionOptions() throws SSLException
     {
-        EncryptionOptions jmxEncryptionOptions = DatabaseDescriptor.getJmxEncryptionOptions();
-        String expectedProtocols = StringUtils.join(jmxEncryptionOptions.getAcceptedProtocols(), ",");
-        String expectedCipherSuites = StringUtils.join(jmxEncryptionOptions.cipherSuitesArray(), ",");
-
         InetAddress serverAddress = InetAddress.getLoopbackAddress();
         COM_SUN_MANAGEMENT_JMXREMOTE_SSL.setBoolean(false);
         Map<String, Object> env = JMXServerUtils.configureJmxSocketFactories(serverAddress, false);
+        Assert.assertTrue("no properties must be set", env.isEmpty());
+        Assert.assertFalse("com.sun.management.jmxremote.ssl must be false", COM_SUN_MANAGEMENT_JMXREMOTE_SSL.getBoolean());
+        Assert.assertNull("javax.rmi.ssl.client.enabledProtocols must be null", JAVAX_RMI_SSL_CLIENT_ENABLED_PROTOCOLS.getString());
+        Assert.assertNull("javax.rmi.ssl.client.enabledCipherSuites must be null", JAVAX_RMI_SSL_CLIENT_ENABLED_CIPHER_SUITES.getString());
+    }
 
-        Assert.assertTrue("com.sun.management.jmxremote.ssl must be true", COM_SUN_MANAGEMENT_JMXREMOTE_SSL.getBoolean());
+    /**
+     * Tests fallback to the `local only` JMX server when jmx_encryption_options are disabled in the cassandra.yaml
+     * and no System settings provided for the remote SSL config.
+     * @throws SSLException
+     */
+    @Test
+    public void testJMXSslConfigFallbackToLocal() throws SSLException
+    {
+        InetAddress serverAddress = InetAddress.getLoopbackAddress();
+        COM_SUN_MANAGEMENT_JMXREMOTE_SSL.setBoolean(false);
+        Map<String, Object> env = JMXServerUtils.configureJmxSocketFactories(serverAddress, true);
+        Assert.assertFalse("com.sun.management.jmxremote.ssl must be false", COM_SUN_MANAGEMENT_JMXREMOTE_SSL.getBoolean());
+        Assert.assertNull("com.sun.jndi.rmi.factory.socket must be null", env.get("com.sun.jndi.rmi.factory.socket"));
         Assert.assertNotNull("ServerSocketFactory must not be null", env.get(RMIConnectorServer.RMI_SERVER_SOCKET_FACTORY_ATTRIBUTE));
-        Assert.assertNotNull("ClientSocketFactory must not be null", env.get(RMIConnectorServer.RMI_CLIENT_SOCKET_FACTORY_ATTRIBUTE));
-        Assert.assertNotNull("com.sun.jndi.rmi.factory.socket must be set in the env", env.get("com.sun.jndi.rmi.factory.socket"));
-        Assert.assertEquals("javax.rmi.ssl.client.enabledProtocols must match", expectedProtocols, JAVAX_RMI_SSL_CLIENT_ENABLED_PROTOCOLS.getString());
-        Assert.assertEquals("javax.rmi.ssl.client.enabledCipherSuites must match", expectedCipherSuites, JAVAX_RMI_SSL_CLIENT_ENABLED_CIPHER_SUITES.getString() );
+        Assert.assertFalse("com.sun.management.jmxremote.ssl must be false", COM_SUN_MANAGEMENT_JMXREMOTE_SSL.getBoolean());
+        Assert.assertNull("javax.rmi.ssl.client.enabledProtocols must be null", JAVAX_RMI_SSL_CLIENT_ENABLED_PROTOCOLS.getString());
+        Assert.assertNull("javax.rmi.ssl.client.enabledCipherSuites must be null", JAVAX_RMI_SSL_CLIENT_ENABLED_CIPHER_SUITES.getString());
     }
 }
